refactor(frontend): config-io の forms 依存を逆転、render/collect を fan-out 化
- forms/* (timestamp/search/folders/apps) に collect() を新規 export
- lib/config-io.js は forms/* を import せず、initConfigIo({ renderers, collectors }) で受け取る
- main.js で render/collect を集約して config-io に注入
- 動作変化なし、collectConfig の戻り値構造と呼び出し順は従来通り
- punctuation_style の DOM 操作のみ config-io に残置 (Issue #140 受け入れ基準は充足、追い出しは別 Issue 候補)
- Phase 3 TS 化と Phase 4 単体テストの布石

Closes #140
